retry task 100

+++ 0x1E-search_algorithms/100-jump.c
@@ -10,7 +10,7 @@
 int jump_search(int *array, size_t size, int value)
 {
 	size_t step = 0;
-	int i = 0;
+	size_t i = 0;
 	size_t x;
 
 	if (size < 1)
@@ -24,7 +24,7 @@ int jump_search(int *array, size_t size, int value)
 
 	if (array[i] <= value)
 	{
-		printf("Value found between indexes [%d] and [%lu]\n", i, step);
+		printf("Value found between indexes [%lu] and [%lu]\n", i, step);
 		for (x = i; x <= step && x < size; x++)
 		{
 			printf("Value checked array[%lu] = [%d]\n", x, array[x]);
